validate agg func in doValidateInternal
